requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
PyPDF2
faiss-cpu
pandas
python-dotenv